+++ features/step_definitions/record/common.ts
@@ -207,11 +207,12 @@ When("I type {string}", function (userInput) {
 
 When("I press Enter", function () {
   return new Promise<void>((resolve) => {
-    this.childProcess.stdin.write("\n");
-    this.childProcess.stdin.end();
     this.childProcess.on("close", () => {
       resolve();
     });
+
+    this.childProcess.stdin.write("\n");
+    this.childProcess.stdin.end();
   });
 });