+++ src/__tests__/setup.ts
@@ -1,3 +1,4 @@
+import { beforeEach } from "vitest";
 import { epic } from "allure-js-commons";
 
 beforeEach(() => {