+++ src/record/import/usecases/__tests__/upsert/fixtures/upsertByRecordNumberWithAppCodeOnKintone/index.ts
@@ -3,14 +3,12 @@ import { records } from "./records";
 import { schema } from "./schema";
 import { expected } from "./expected";
 import { recordsOnKintone } from "./recordsOnKintone";
-import { LocalRecordRepositoryMock } from "../../../../../repositories/localRecordRepositoryMock";
 
 export const pattern: TestPattern = {
   description:
     "should upsert records correctly with record number on kintone has app code",
   input: {
-    records: records,
-    repository: new LocalRecordRepositoryMock(records, "csv"),
+    repository: { source: records, format: "csv" },
     schema: schema,
     updateKey: "recordNumber",
     options: {