+++ src/record/import/usecases/__tests__/upsert/fixtures/upsertWithMissingKeyFromRecord/index.ts
@@ -2,14 +2,12 @@ import type { TestPattern } from "../../index.test";
 import { records } from "./records";
 import { schema } from "../schema";
 import { recordsOnKintone } from "../recordsOnKintone";
-import { LocalRecordRepositoryMock } from "../../../../../repositories/localRecordRepositoryMock";
 
 export const pattern: TestPattern = {
   description:
     "should throw error when update key field does not exist on input record",
   input: {
-    records: records,
-    repository: new LocalRecordRepositoryMock(records, "csv"),
+    repository: { source: records, format: "csv" },
     schema: schema,
     updateKey: "singleLineText_nonExistentOnInput",
     options: {