+++ website/docusaurus.config.ts
@@ -40,7 +40,7 @@ const config: Config = {
         label: "English",
       },
       ja: {
-        label: "日本語 (29%)",
+        label: "日本語 (34%)",
       },
     },
   },